requirements.txt
python-dotenv
django-tinymce
validate_email
Pillow
django
mysqlclient